+++ zcb_app/sms/service/dao/src/main/java/com/zcb_app/sms/service/dao/type/MsgVerifyCodeParams.java
@@ -1,14 +1,12 @@
 package com.zcb_app.sms.service.dao.type;
 
-import java.math.BigInteger;
-
 import com.zcb_app.sms.service.facade.dataobject.VerifyCodeParams;
 
 public class MsgVerifyCodeParams {
 
 	private String mobile;//手机号
 	
-	private BigInteger tmpl_id;//短信模板ID
+	private Long tmpl_id;//短信模板ID
 	
 	private String relation_key;//关联的参数，下发时传入，验证时必须传入
 	
@@ -24,11 +22,11 @@ public class MsgVerifyCodeParams {
 		this.mobile = mobile;
 	}
 
-	public BigInteger getTmpl_id() {
+	public Long getTmpl_id() {
 		return tmpl_id;
 	}
 
-	public void setTmpl_id(BigInteger tmpl_id) {
+	public void setTmpl_id(Long tmpl_id) {
 		this.tmpl_id = tmpl_id;
 	}
