+++ zcb_app/sms/service/facade/src/main/java/com/zcb_app/sms/service/facade/dataobject/MobileLimitDO.java
@@ -1,7 +1,6 @@
 package com.zcb_app.sms.service.facade.dataobject;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.Date;
 
 
@@ -14,21 +13,21 @@ public class MobileLimitDO implements Serializable {
 	
 	private static final long serialVersionUID = 4726699177919005728L;
 
-	private BigInteger fpkid;//流水编号（自增主键）
+	private long fpkid;//流水编号（自增主键）
 	
 	private String fmobile_no;//手机号码
 	
-	private BigInteger ftmpl_id;//短信模板id，触发受限的短信模板id
+	private Long ftmpl_id;//短信模板id，触发受限的短信模板id
 	
 	private Date fblock_stime;//受限制的开始时间
 	
-	private BigInteger fblock_timespan;//受限制的时长，秒为单位，只有当前时间大于fblock_stime+ fblock_timespan才解除限制
+	private long fblock_timespan;//受限制的时长，秒为单位，只有当前时间大于fblock_stime+ fblock_timespan才解除限制
 
-	public BigInteger getFpkid() {
+	public long getFpkid() {
 		return fpkid;
 	}
 
-	public void setFpkid(BigInteger fpkid) {
+	public void setFpkid(long fpkid) {
 		this.fpkid = fpkid;
 	}
 
@@ -40,11 +39,11 @@ public class MobileLimitDO implements Serializable {
 		this.fmobile_no = fmobile_no;
 	}
 
-	public BigInteger getFtmpl_id() {
+	public Long getFtmpl_id() {
 		return ftmpl_id;
 	}
 
-	public void setFtmpl_id(BigInteger ftmpl_id) {
+	public void setFtmpl_id(Long ftmpl_id) {
 		this.ftmpl_id = ftmpl_id;
 	}
 
@@ -56,11 +55,11 @@ public class MobileLimitDO implements Serializable {
 		this.fblock_stime = fblock_stime;
 	}
 
-	public BigInteger getFblock_timespan() {
+	public long getFblock_timespan() {
 		return fblock_timespan;
 	}
 
-	public void setFblock_timespan(BigInteger fblock_timespan) {
+	public void setFblock_timespan(long fblock_timespan) {
 		this.fblock_timespan = fblock_timespan;
 	}
