+++ zcb_app/sms/service/facade/src/main/java/com/zcb_app/sms/service/facade/dataobject/VerifyCodeParams.java
@@ -1,7 +1,6 @@
 package com.zcb_app.sms.service.facade.dataobject;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 public class VerifyCodeParams implements Serializable {
 	
@@ -9,7 +8,7 @@ public class VerifyCodeParams implements Serializable {
 
 	private String mobile;//手机号
 	
-	private BigInteger tmpl_id;//短信模板ID
+	private Long tmpl_id;//短信模板ID
 	
 	private String verify_code;//验证码
 	
@@ -25,11 +24,11 @@ public class VerifyCodeParams implements Serializable {
 		this.mobile = mobile;
 	}
 
-	public BigInteger getTmpl_id() {
+	public Long getTmpl_id() {
 		return tmpl_id;
 	}
 
-	public void setTmpl_id(BigInteger tmpl_id) {
+	public void setTmpl_id(Long tmpl_id) {
 		this.tmpl_id = tmpl_id;
 	}
